+++ src/client/pythonEnvironments/common/environmentIdentifier.ts
@@ -4,7 +4,7 @@
 import { PythonEnvKind } from '../base/info';
 import { getPrioritizedEnvKinds } from '../base/info/envKind';
 import { isCondaEnvironment } from './environmentManagers/conda';
-import { isPipenvEnvironment } from '../discovery/locators/services/pipenv';
+import { isPipenvEnvironment } from './environmentManagers/pipenv';
 import { isPoetryEnvironment } from './environmentManagers/poetry';
 import { isPyenvEnvironment } from '../discovery/locators/services/pyenvLocator';
 import {
